src/requirements.txt
pandas
numpy
scikit-learn
pyyaml
uvicorn
dvc
dvc-gdrive 